晚上日常提交

+++ src/main/java/cn/hgxsp/miaosha_1/Domain/OrderInfo.java
@@ -5,10 +5,10 @@ import lombok.Data;
 import java.util.Date;
 @Data
 public class OrderInfo {
-	private Long id;
-	private Long userId;
-	private Long goodsId;
-	private Long  deliveryAddrId;
+	private int id;
+	private int userId;
+	private int goodsId;
+	private int  deliveryAddrId;
 	private String goodsName;
 	private Integer goodsCount;
 	private Double goodsPrice;
